src/gear_gemini_agent.py
import os
import google.generativeai as genai
from typing import Dict, Any, List
from .gear_adk_base import GEARBaseAgent
from dotenv import load_dotenv

load_dotenv()

class GeminiReasoningAgent(GEARBaseAgent):
    """
    GEAR Agent that leverages Gemini 1.5 Pro for deep reasoning
    on detected anomalies.
    """
    def __init__(self, agent_id: str = "GEMINI_REASONER_AI"):
        super().__init__(agent_id)
        api_key = os.getenv("GOOGLE_API_KEY")
        if api_key:
            genai.configure(api_key=api_key)
            self.model = genai.GenerativeModel('gemini-1.5-pro')
            self.active = True
        else:
            self.log("GOOGLE_API_KEY not found. Gemini reasoning disabled.", "ERROR")
            self.active = False

    def process(self, anomaly_data: Dict[str, Any]) -> str:
        """
        Uses Gemini to reason about a specific anomaly.
        """
        if not self.active:
            return "Reasoning unavailable (No API Key)"

        prompt = f"""
        Analyze the following aviation telemetry anomaly for potential cybersecurity threats (ADS-B Spoofing or ARINC Injection).
        
        DATA:
        {anomaly_data}
        
        TASK:
        1. Identify if this matches known spoofing patterns.
        2. Explain the physical possibility/impossibility of the jump.
        3. Provide a 'Confidence Score' (0-100).
        4. Suggest a mitigation action.
        
        Format your response as a concise forensic summary.
        """
        
        try:
            response = self.model.generate_content(prompt)
            return response.text
        except Exception as e:
            self.log(f"Gemini processing error: {e}", "CRITICAL")
            return f"Error in reasoning: {str(e)}"

    def audit_flight_block(self, telemetry_block: List[Dict[str, Any]]):
        """Batch auditing for Big Data sets."""
        self.log(f"Starting mass audit of {len(telemetry_block)} records via Gemini Pro.", "GEAR_SECURITY")
        # Simplified batch logic
        pass
