requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
pytest
streamlit
google-cloud-bigquery
google-api-core
tensorflow
google-generative-ai
pydantic
python-dotenv
